Update wording in README
It was fairly imprecise before, this is not really property-based
testing because there is no automated process testing the property. It's
intended to show any issues with potential errors or weird values that
would be visible to the tester themselves (and they may not have written
a test to cover that case yet as they haven't thought about it).